Moved Knowledge panel + children out of KnowledgePanel folder
